Refactor dice and bonus extraction logic tests
